Functional Firestore data converter
